removed useState and excess copy pasted code for edit functionality in ActivityEndForm.js
